+++ .build/tasks/DscResource.Test.build.ps1
@@ -15,10 +15,6 @@ Param (
 
     [Parameter()]
     [string]
-    $ModuleVersion = (property ModuleVersion ''),
-
-    [Parameter()]
-    [string]
     $DscTestOutputFolder = (property DscTestOutputFolder 'testResults'),
 
     [Parameter()]
@@ -63,12 +59,11 @@ task Invoke_DscResource_tests {
     }
 
     $getModuleVersionParameters = @{
-        OutputDirectory = $BuildModuleOutput
+        OutputDirectory = $OutputDirectory
         ProjectName     = $ProjectName
-        ModuleVersion   = $ModuleVersion
     }
 
-    $ModuleVersion = Get-ModuleVersion @getModuleVersionParameters
+    $ModuleVersion = Get-BuiltModuleVersion @getModuleVersionParameters
 
     if (!(Test-Path $DscTestOutputFolder))
     {
@@ -241,10 +236,9 @@ task Fail_Build_if_DscResource_Tests_failed {
     $getModuleVersionParameters = @{
         OutputDirectory = $OutputDirectory
         ProjectName     = $ProjectName
-        ModuleVersion   = $ModuleVersion
     }
 
-    $ModuleVersion = Get-ModuleVersion @getModuleVersionParameters
+    $ModuleVersion = Get-BuiltModuleVersion @getModuleVersionParameters
 
     $PSVersion = 'PSv.{0}' -f $PSVersionTable.PSVersion
     $DscTestOutputFileFileName = "DscTest_{0}_v{1}.{2}.{3}.xml" -f $ProjectName, $ModuleVersion, $os, $PSVersion
@@ -304,10 +298,9 @@ task Upload_DscResourceTest_Results_To_AppVeyor -If { (property BuildSystem 'unk
     $getModuleVersionParameters = @{
         OutputDirectory = $OutputDirectory
         ProjectName     = $ProjectName
-        ModuleVersion   = $ModuleVersion
     }
 
-    $ModuleVersion = Get-ModuleVersion @getModuleVersionParameters
+    $ModuleVersion = Get-BuiltModuleVersion @getModuleVersionParameters
 
     $PSVersion = 'PSv.{0}' -f $PSVersionTable.PSVersion
     $DscTestOutputFileFileName = "DscResource.Test_{0}_v{1}.{2}.{3}.xml" -f $ProjectName, $ModuleVersion, $os, $PSVersion